[clang] Fix -Wformat warnings
Use %hx for uint16_t instead of %hhx.

+++ app/stringtests/string_tests.c
@@ -267,7 +267,7 @@ static void validate_memset(void) {
 
                 int comp = memcmp(dst, dst2, maxsize * 2);
                 if (comp != 0) {
-                    printf("error! align %zu, c 0x%hhx, size %zu\n", dstalign, c, size);
+                    printf("error! align %zu, c 0x%hhx, size %zu\n", dstalign, (uint8_t)c, size);
                 }
             }
         }
